requirements.txt
Flask
azure-cognitiveservices-vision-computervision
firebase-admin
gunicorn